+++ src/mako/pixel/image/ColorFamily.php
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\pixel\image;
+
+/**
+ * Color family.
+ */
+enum ColorFamily: string
+{
+	case Red = 'red';
+	case Orange = 'orange';
+	case Yellow = 'yellow';
+	case Green = 'green';
+	case Cyan = 'cyan';
+	case Blue = 'blue';
+	case Purple = 'purple';
+	case Pink = 'pink';
+	case Brown = 'brown';
+	case Black = 'black';
+	case Gray = 'gray';
+	case White = 'white';
+}