+++ tests/unit/pixel/image/ColorTest.php
@@ -9,6 +9,7 @@ namespace mako\tests\unit\pixel\image;
 
 use InvalidArgumentException;
 use mako\pixel\image\Color;
+use mako\pixel\image\ColorFamily;
 use mako\pixel\image\WebColor;
 use mako\tests\TestCase;
 use PHPUnit\Framework\Attributes\Group;
@@ -418,5 +419,130 @@ class ColorTest extends TestCase
 		$color = new Color(255, 0, 0, 127);
 
 		$this->assertSame('hwb(0 0.0% 0.0% / 0.498)', $color->toHwbaString());
+	}
+
+	/**
+	 *
+	 */
+	public function testToColorFamily(): void
+	{
+		// Key = Color, Value = Expected Family
+
+		$colors = [
+			// Basic colors
+
+			WebColor::Red->value => ColorFamily::Red,
+			WebColor::Orange->value => ColorFamily::Orange,
+			WebColor::Yellow->value => ColorFamily::Yellow,
+			WebColor::Green->value => ColorFamily::Green,
+			WebColor::Cyan->value => ColorFamily::Cyan,
+			WebColor::Blue->value => ColorFamily::Blue,
+			WebColor::Purple->value => ColorFamily::Purple,
+			WebColor::Pink->value => ColorFamily::Pink,
+			WebColor::Brown->value => ColorFamily::Brown,
+			WebColor::Black->value => ColorFamily::Black,
+			WebColor::Gray->value => ColorFamily::Gray,
+			WebColor::White->value => ColorFamily::White,
+
+			// Red
+
+			WebColor::DarkRed->value => ColorFamily::Red,
+			WebColor::Crimson->value => ColorFamily::Red,
+			WebColor::IndianRed->value => ColorFamily::Red,
+
+			// Orange
+
+			WebColor::OrangeRed->value => ColorFamily::Orange,
+			WebColor::DarkOrange->value => ColorFamily::Orange,
+			WebColor::Coral->value => ColorFamily::Orange,
+
+			// Yellow
+
+			WebColor::Gold->value => ColorFamily::Yellow,
+			WebColor::PaleGoldenrod->value => ColorFamily::Yellow,
+			WebColor::LightYellow->value => ColorFamily::Yellow,
+			WebColor::LightGoldenrodYellow->value => ColorFamily::Yellow,
+
+			// Green
+
+			WebColor::DarkGreen->value => ColorFamily::Green,
+			WebColor::ForestGreen->value => ColorFamily::Green,
+			WebColor::SeaGreen->value => ColorFamily::Green,
+			WebColor::MediumSeaGreen->value => ColorFamily::Green,
+			WebColor::LimeGreen->value => ColorFamily::Green,
+			WebColor::SpringGreen->value => ColorFamily::Green,
+			WebColor::MediumSpringGreen->value => ColorFamily::Green,
+			WebColor::LawnGreen->value => ColorFamily::Green,
+			WebColor::LightGreen->value => ColorFamily::Green,
+			WebColor::PaleGreen->value => ColorFamily::Green,
+
+			// Cyan
+
+			WebColor::Teal->value => ColorFamily::Cyan,
+			WebColor::DarkCyan->value => ColorFamily::Cyan,
+			WebColor::DarkTurquoise->value => ColorFamily::Cyan,
+			WebColor::Turquoise->value => ColorFamily::Cyan,
+			WebColor::PaleTurquoise->value => ColorFamily::Cyan,
+			WebColor::LightCyan->value => ColorFamily::Cyan,
+
+			// Blue
+
+			WebColor::MidnightBlue->value => ColorFamily::Blue,
+			WebColor::Navy->value => ColorFamily::Blue,
+			WebColor::DarkBlue->value => ColorFamily::Blue,
+			WebColor::MediumBlue->value => ColorFamily::Blue,
+			WebColor::RoyalBlue->value => ColorFamily::Blue,
+			WebColor::SteelBlue->value => ColorFamily::Blue,
+			WebColor::DodgerBlue->value => ColorFamily::Blue,
+			WebColor::DeepSkyBlue->value => ColorFamily::Blue,
+			WebColor::SkyBlue->value => ColorFamily::Blue,
+			WebColor::LightSkyBlue->value => ColorFamily::Blue,
+
+			// Purple
+
+			WebColor::Indigo->value => ColorFamily::Purple,
+			WebColor::DarkMagenta->value => ColorFamily::Purple,
+			WebColor::DarkViolet->value => ColorFamily::Purple,
+			WebColor::BlueViolet->value => ColorFamily::Purple,
+			WebColor::DarkOrchid->value => ColorFamily::Purple,
+			WebColor::MediumOrchid->value => ColorFamily::Purple,
+			WebColor::MediumPurple->value => ColorFamily::Purple,
+			WebColor::Orchid->value => ColorFamily::Purple,
+			WebColor::Violet->value => ColorFamily::Purple,
+			WebColor::Thistle->value => ColorFamily::Purple,
+
+			// Pink
+
+			WebColor::PaleVioletRed->value => ColorFamily::Pink,
+			WebColor::HotPink->value => ColorFamily::Pink,
+			WebColor::LightPink->value => ColorFamily::Pink,
+
+			// Brown
+
+			WebColor::SaddleBrown->value => ColorFamily::Brown,
+			WebColor::Sienna->value => ColorFamily::Brown,
+
+			// Gray
+
+			WebColor::DimGray->value => ColorFamily::Gray,
+			WebColor::DarkGray->value => ColorFamily::Gray,
+			WebColor::Silver->value => ColorFamily::Gray,
+			WebColor::LightGray->value => ColorFamily::Gray,
+			WebColor::Gainsboro->value => ColorFamily::Gray,
+
+			// White
+
+			WebColor::WhiteSmoke->value => ColorFamily::White,
+			WebColor::Seashell->value => ColorFamily::White,
+			WebColor::GhostWhite->value => ColorFamily::White,
+			WebColor::Honeydew->value => ColorFamily::White,
+			WebColor::MintCream->value => ColorFamily::White,
+			WebColor::Snow->value => ColorFamily::White,
+			WebColor::Ivory->value => ColorFamily::White,
+		];
+
+		foreach ($colors as $color => $expected) {
+			$this->assertSame($expected, Color::fromHex($color)->toColorFamily());
+		}
 	}
 }