+++ CHANGELOG.md
@@ -80,6 +80,8 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 	- `Color::fromWebColor()`
 	- `Color::invert()`
 	- `Color::with()`
+	- `Color::toColorFamily()`
+* Added `ColorFamily` enum.
 * Added `WebColor` enum.
 * Added dotenv loader.
 * Added `Signer::validateOrThrow()` method.